requirements.txt
aiohttp==3.9.5
aiolimiter==1.1.0
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
discord.py==2.3.2
Django==5.0.6
et-xmlfile==1.1.0
frozenlist==1.4.1
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
h11==0.14.0
httplib2==0.22.0
idna==3.7
imaplib2==3.6
lxml==5.2.2
MouseInfo==0.1.3
multidict==6.0.5
numpy==2.0.0
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.3.0.post0
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyarmor==8.5.10
pyarmor.cli.core==6.5.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.21.0
simplegmail==4.1.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wsproto==1.2.0
yarl==1.9.4